+++ sdk/spring/spring-cloud-azure-resourcemanager/src/main/java/com/azure/spring/cloud/resourcemanager/implementation/crud/ResourceCrud.java
@@ -6,11 +6,12 @@ package com.azure.spring.cloud.resourcemanager.implementation.crud;
 /**
  * Interface to support CRUD of Azure Resource
  *
- * @param <T> Azure resource type
- * @param <K> Azure resource key type
+ * @param <T> Azure resource type.
+ * @param <K> Azure resource key type.
+ * @param <P> Azure resource properties.
  *
  */
-public interface ResourceCrud<T, K> {
+public interface ResourceCrud<T, K, P> {
 
     /**
      * Gets a value for a given key.
@@ -29,18 +30,20 @@ public interface ResourceCrud<T, K> {
     boolean exists(K key);
 
     /**
-     * Creates a value for a given key.
+     * Creates a value for a given key and properties.
      *
      * @param key The key.
+     * @param properties The properties for the Azure resource creation.
      * @return The created value.
      */
-    T create(K key);
+    T create(K key, P properties);
 
     /**
-     * Gets or creates a value for a given key.
+     * Gets or creates a value for a given key and properties.
      *
      * @param key The key.
+     * @param properties The properties for the Azure resource creation.
      * @return The retrieved or created value.
      */
-    T getOrCreate(K key);
+    T getOrCreate(K key, P properties);
 }