+++ sdk/spring/spring-cloud-azure-resourcemanager/src/main/java/com/azure/spring/cloud/resourcemanager/implementation/crud/ResourceGroupCrud.java
@@ -7,11 +7,12 @@ import com.azure.core.management.exception.ManagementException;
 import com.azure.resourcemanager.AzureResourceManager;
 import com.azure.resourcemanager.resources.models.ResourceGroup;
 import com.azure.spring.cloud.core.properties.resource.AzureResourceMetadata;
+import org.springframework.lang.Nullable;
 
 /**
  * Resource manager for resource group.
  */
-public class ResourceGroupCrud extends AbstractResourceCrud<ResourceGroup, String> {
+public class ResourceGroupCrud extends AbstractResourceCrud<ResourceGroup, String, Object> {
 
     /**
      * Creates a new instance of {@link ResourceGroupCrud}.
@@ -47,7 +48,7 @@ public class ResourceGroupCrud extends AbstractResourceCrud<ResourceGroup, Strin
     }
 
     @Override
-    public ResourceGroup internalCreate(String key) {
+    public ResourceGroup internalCreate(String key, @Nullable Object properties) {
         return resourceManager.resourceGroups()
                               .define(key)
                               .withRegion(resourceMetadata.getResourceGroup())